Computing non-parametric mixing distribution

Computing non-parametric mixing distribution

computemixdist(x, ...)

Arguments

x
a object from implemented family generated by makeobject.

...
parameters above passed to the specific method

Details

The full list of implemented family is in makeobject.

The avaliable parameters are listed as follows:

mix: The initial proper mixing distribution

tol: tolerance to stop the code

maxiter: maximum iterations allowed.

verbose: logical; whether to print the intermediate results.

This function essentially calls the class method in the object.

Examples

data = rnorm(500, c(0, 2))
pi0 = 0.5
x = makeobject(data, pi0 = pi0, method = "npnormll")
computemixdist(x)

#> $iter
#> [1] 14
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -1.257065e-06
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.789346 2.701833
#> 
#> $mix$pr
#> [1] 0.5000000 0.3758146 0.1241854
#> 
#> 
#> $ll
#> [1] 890.1905
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, pi0 = pi0, method = "npnormllw")
computemixdist(x)

#> $iter
#> [1] 14
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -1.120656e-06
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.789414 2.701783
#> 
#> $mix$pr
#> [1] 0.5000000 0.3758313 0.1241687
#> 
#> 
#> $ll
#> [1] 890.193
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, pi0 = pi0, method = "npnormcvm")
computemixdist(x)

#> $iter
#> [1] 11
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -1.496881e-06
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.987451 4.669628
#> 
#> $mix$pr
#> [1] 0.5000000 0.4859121 0.0140879
#> 
#> 
#> $ll
#> [1] 0.04815006
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, pi0 = pi0, method = "npnormcvmw")
computemixdist(x)

#> $iter
#> [1] 10
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -6.794002e-06
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.988785 4.704250
#> 
#> $mix$pr
#> [1] 0.50000000 0.48614493 0.01385507
#> 
#> 
#> $ll
#> [1] 0.0482996
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, pi0 = pi0, method = "npnormad")
computemixdist(x)

#> $iter
#> [1] 11
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -2.469989e-08
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.969779 3.640038
#> 
#> $mix$pr
#> [1] 0.50000000 0.47508101 0.02491899
#> 
#> 
#> $ll
#> [1] 500.3006
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, pi0 = pi0, method = "nptll")
computemixdist(x)

#> $iter
#> [1] 9
#> 
#> $family
#> [1] "npt"
#> 
#> $max.gradient
#> [1] -1.423797e-06
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.786465 2.696622
#> 
#> $mix$pr
#> [1] 0.5000000 0.3741084 0.1258916
#> 
#> 
#> $ll
#> [1] 890.1905
#> 
#> $beta
#> [1] Inf
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"