Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(x, ...)

Arguments

x
a object from implemented family

...
parameters above passed to the specific method.

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

val: Threshold value

mix: The initial proper mixing distribution.

tol: tolerance to stop the code.

maxiter: maximum iterations allowed.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default thresold values and this function essentially calls the class method in the object.

The full list of implemented families is in makeobject.

Examples

data = rnorm(500, c(0, 2))
pi0 = 0.5
x = makeobject(data, method = "npnormll")
estpi0(x)

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -6.41685e-06
#> 
#> $mix
#> $mix$pt
#> [1] -1.452779  0.000000  1.970093
#> 
#> $mix$pr
#> [1] 0.001175823 0.465355758 0.533468419
#> 
#> 
#> $ll
#> [1] 876.598
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, method = "npnormllw")
estpi0(x)

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] -6.401832e-06
#> 
#> $mix
#> $mix$pt
#> [1] -1.452738  0.000000  1.970067
#> 
#> $mix$pr
#> [1] 0.00117343 0.46535697 0.53346960
#> 
#> 
#> $ll
#> [1] 876.5945
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, method = "npnormcvm")
estpi0(x)

#> $iter
#> [1] 5
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] 0
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.085546
#> 
#> $mix$pr
#> [1] 0.4936154 0.5063846
#> 
#> 
#> $ll
#> [1] 0.1187139
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, method = "npnormcvm")
estpi0(x)

#> $iter
#> [1] 5
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] 0
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.085546
#> 
#> $mix$pr
#> [1] 0.4936154 0.5063846
#> 
#> 
#> $ll
#> [1] 0.1187139
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, method = "npnormad")
estpi0(x)

#> $iter
#> [1] 15
#> 
#> $family
#> [1] "npnorm"
#> 
#> $max.gradient
#> [1] 5.684342e-14
#> 
#> $mix
#> $mix$pt
#> [1] 0.00000 2.07922
#> 
#> $mix$pr
#> [1] 0.4978127 0.5021873
#> 
#> 
#> $ll
#> [1] 500.7742
#> 
#> $beta
#> [1] 1
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

x = makeobject(data, method = "nptll")
estpi0(x)

#> Error in dt(x, ncp = rep(mu0, rep(length(x), length(mu0))), df = df): Non-numeric argument to mathematical function