Making object for computation

These functions are used to make the object for computing the non-paramtric mixing distribution or estimating the proportion of zero using non-parametric methods.

makeobject(v, method = "npnormll", ...)

makeobject.npnormad(v, mu0, pi0, beta, order)

makeobject.npnormadw(v, mu0, pi0, beta, order = -3)

makeobject.npnormcll(v, mu0, pi0, beta, order)

makeobject.npnormcvm(v, mu0, pi0, beta, order)

makeobject.npnormcvmw(v, mu0, pi0, beta, order = -3)

makeobject.npnormll(v, mu0, pi0, beta, order)

makeobject.npnormllw(v, mu0, pi0, beta, order = -3)

makeobject.nptll(v, mu0, pi0, beta, order)

Arguments

v
the object either numeric or the implmented family

method
An implemented family; see details

...
other parameter passed to the constructor.

mu0
A vector of support points

pi0
A vector of weights corresponding to the support points

beta
structual parameter.

order
the parameter for the binned version.

Details

This is a generic function for making the object for computing the non-parametric mixing distribution or estimating the proportion of zero.

current implemented families are:

npnormll : normal density using maximum likelihood (Chapter 3). The default beta is 1.

npnormllw : Binned version of normal density using maximum likelihood (Chapter 6). The default beta is 1 and the default order is -3.

npnormcvm : normal density using the Cramer-von Mises distance (Chapter 5). The default beta is 1.

npnormcvmw : Binned version of normal density using the Cramer-von Mises distance (Chapter 6). The default beta is 1 and the default order is -3.

npnormad : normal density using the Anderson-Darling distance (Chapter 5). The default beta is 1.

npnormadw : Binned version of normal density using the Anderson-Darling distance (Chapter 6) The default beta is 1 and the default order is -3.

nptll : t-density using maximum likelihood (Chapter 3). The default beta is infinity (normal distribution).

npnormcll : the one-parameter normal distribution used for approximating the sample correlation coefficients using maximum likelihood. This does not have a corresponding estimation of zero due to incompleted theory (Chapter 8). There is no default beta. The structure beta is the number of observations.

The default method used is npnormll.

The detailed description of the npfixedcompRobject class is in npfixedcompRobject.