Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
x = makeobject(data, pi0 = 0.5)
r1 = computemixdist(x)
posteriormean(data, r1)

#>   [1] -6.813116e-02  1.215570e+00  1.273850e-01  1.974569e+00  1.953525e-01
#>   [6]  4.550760e-01  1.729214e-01  1.737402e+00  1.528984e-01  2.056186e+00
#>  [11]  1.224241e-01  2.244701e+00  1.836643e-01  2.186652e+00  1.000861e+00
#>  [16]  9.682113e-01 -1.199688e-01  1.787875e+00  2.959946e-02  1.987431e+00
#>  [21]  1.743898e+00  5.909183e-01  5.016702e-01  2.023493e+00 -6.358579e-02
#>  [26]  1.128813e+00  6.411676e-02  1.862926e+00  2.984376e-01  2.084402e+00
#>  [31] -6.842327e-03  8.431061e-01  2.423542e-01  1.908940e+00 -8.002007e-02
#>  [36]  1.983031e+00  7.440792e-01  2.072389e+00 -5.481187e-01  1.755414e+00
#>  [41]  5.770983e-02  2.030553e+00 -1.861477e-01  2.032160e+00  4.541282e-02
#>  [46]  3.945831e-01 -3.100278e-01  2.168071e+00  1.193103e+00  2.041796e+00
#>  [51]  8.338975e-01  2.001647e+00 -4.023913e-02  2.004959e+00  9.132337e-01
#>  [56]  8.244954e-01 -1.954452e-01  1.945884e+00  7.583920e-01  1.887474e+00
#>  [61]  1.020229e-01  1.631498e+00  2.782549e-01  2.050610e+00  1.821906e+00
#>  [66]  1.863176e+00  1.035337e+00  2.065004e+00 -1.080939e-01  1.766286e+00
#>  [71]  5.534417e-01  1.905429e+00  1.177692e-01  1.936501e+00  1.138806e-01
#>  [76]  1.257041e+00 -5.790964e-02  6.111953e-01 -1.283847e-01  1.308339e+00
#>  [81]  2.562183e-01  1.357615e+00  1.253696e-02  2.063423e+00 -3.563531e-01
#>  [86]  2.281793e+00  1.515603e+00  1.153911e+00  3.510133e-01  1.468049e+00
#>  [91]  2.607969e-02  1.736832e+00  2.072792e-01  1.593477e+00  1.376406e+00
#>  [96]  1.838155e+00  1.819465e-02  8.765841e-01  1.145609e+00  3.298815e-01
#> [101]  7.912040e-02  1.904279e+00 -4.882454e-01  1.943388e+00  2.581053e-01
#> [106]  1.608517e+00  4.354615e-01  2.062067e+00 -1.511403e-01  1.968532e+00
#> [111]  7.941926e-01  1.492758e+00  5.756824e-01  3.419616e-01  7.303805e-01
#> [116]  1.653904e+00  2.737056e-01  2.180795e+00 -1.900233e-01  1.735835e+00
#> [121] -5.928372e-03  2.205548e+00 -2.265618e-01  1.939413e+00  2.986597e-01
#> [126]  1.620857e+00 -1.049037e-01  1.915378e+00 -2.080188e-01  1.991953e+00
#> [131] -1.013805e-01  2.077295e+00  8.903722e-01  7.460349e-01  1.778547e+00
#> [136]  1.897336e+00  1.078415e+00  4.897860e-01  1.477209e+00  2.007202e+00
#> [141]  1.042460e-01  1.297570e+00 -5.372033e-01  9.306239e-01  2.013991e+00
#> [146]  1.691712e+00  2.772856e-02  1.567002e+00  5.382842e-01  1.378350e+00
#> [151]  2.644823e-02  2.151419e+00  3.022595e-03  1.494321e+00  3.035304e-01
#> [156]  2.042331e+00  8.310102e-01  2.034999e+00 -1.397501e-01  1.531199e+00
#> [161] -3.399196e-01  1.019682e+00  1.527002e+00  1.922760e+00  1.044347e-01
#> [166]  6.519196e-01  3.441993e-02  2.580959e-01  1.115649e-01  2.009595e+00
#> [171] -1.637833e-02  2.080463e+00 -4.937086e-02  6.571956e-01 -3.150467e-01
#> [176]  1.234182e+00  1.649257e-01  1.454779e+00  1.691807e+00  2.038316e+00
#> [181]  1.016436e+00  2.076806e+00  5.176929e-02  3.257739e+00 -1.844160e-01
#> [186]  1.588680e+00 -2.567949e-01  1.734735e+00 -8.888399e-02  2.104597e+00
#> [191]  1.186109e-01  5.197940e-01  1.272951e-01  2.042440e+00  1.478684e-01
#> [196]  1.110354e+00  3.621777e-01  2.580715e+00  1.026555e+00  2.014764e+00
#> [201]  2.958455e-01  1.848907e+00  9.376551e-01  1.681241e+00  1.785598e+00
#> [206]  1.818197e+00  5.451443e-01  2.055716e+00 -1.720556e-01  2.052515e+00
#> [211]  1.303634e-01  2.746973e-01 -2.598673e-02  2.082208e+00  1.795260e+00
#> [216]  1.431511e+00  4.644566e-01  1.630977e+00 -4.039974e-02  1.055207e+00
#> [221]  5.468872e-01  2.073446e+00  3.991839e-01  2.033637e+00 -4.173885e-02
#> [226]  1.319707e+00 -6.880532e-02  2.150368e+00  2.076232e+00  1.951837e+00
#> [231]  5.253100e-01  1.872983e+00 -9.044565e-02  2.096533e+00  1.829597e-01
#> [236]  2.058110e+00 -3.425009e-02  2.081987e+00  8.277325e-01  2.162542e+00
#> [241]  1.818115e-01  2.044447e+00  4.122696e-02  1.967838e+00  3.276520e-01
#> [246]  1.190371e+00 -6.559827e-01  2.007913e+00  8.413146e-02  1.874381e+00
#> [251] -9.267096e-02  2.136465e+00  2.673258e-01  2.128593e+00 -4.712848e-01
#> [256]  2.079501e+00  1.136873e+00  1.486786e+00  1.056869e+00  1.423410e+00
#> [261]  4.441954e-01  2.132503e+00  1.732913e+00  1.518343e+00  1.861170e+00
#> [266]  1.271172e+00  5.705475e-02  2.070761e+00  1.007367e-01  1.935927e+00
#> [271] -1.122900e-01  1.961282e+00 -3.084634e-01  8.398221e-01 -2.664748e-01
#> [276]  1.089000e+00  5.046318e-02  2.245462e+00  5.331397e-01  5.350420e-01
#> [281]  1.390437e-01  1.480374e+00 -2.319440e-01  1.768265e+00  9.341000e-02
#> [286]  1.134593e+00 -2.903159e-01  7.806962e-01  2.213683e-01  7.969303e-01
#> [291] -3.086847e-01  1.203118e+00 -1.551401e-01  2.029317e+00  2.466156e-01
#> [296]  1.691935e+00 -5.975487e-03  9.380813e-01  9.943995e-01  2.061394e+00
#> [301] -1.072793e-01  2.287891e+00  7.082936e-02  2.910396e+00  3.472476e-01
#> [306]  1.989409e+00  4.124418e-02  1.910704e+00  8.819191e-01  9.683567e-01
#> [311]  4.950154e-01  1.367956e+00  2.188847e-01  1.344899e-01 -5.414330e-02
#> [316]  2.124669e+00  1.245474e-01  1.985301e+00 -8.096316e-02  1.797049e+00
#> [321] -5.890920e-02  6.143993e-01  1.465041e+00  1.275204e+00 -3.787275e-02
#> [326]  9.945652e-01 -1.106382e-01  2.124958e+00  1.745010e-01  8.041828e-01
#> [331]  3.124913e-01  1.849480e+00 -7.768913e-03  2.053728e+00 -2.762493e-02
#> [336]  2.054575e+00  1.089743e+00  2.034554e+00  4.558577e-02  2.095447e+00
#> [341] -7.249315e-02  1.724769e+00 -2.960905e-02  2.508668e-01  1.186010e+00
#> [346]  1.852574e+00  1.006080e-01  1.412618e+00  1.146993e+00  1.913894e+00
#> [351] -2.194696e-01  1.322596e+00 -3.543562e-02  1.990831e+00  6.004080e-01
#> [356]  2.437242e+00  1.889627e+00  2.353000e+00  1.273492e+00  1.573443e+00
#> [361]  1.308883e-01  1.342411e+00 -2.902764e-02  1.632893e+00  4.649701e-05
#> [366]  1.260467e+00  1.465166e+00  2.031798e+00 -2.118421e-02  3.821212e-01
#> [371]  1.123550e+00  2.000933e+00  2.068610e-01  2.087075e+00  9.313250e-01
#> [376]  1.743410e+00 -9.041168e-02  1.954256e+00  4.326270e-01  2.150379e+00
#> [381]  5.870763e-01  1.363469e+00  6.392553e-01  2.073096e+00  1.219475e+00
#> [386]  1.928770e+00  1.358981e-01  1.800395e+00 -2.654306e-02  2.053128e-01
#> [391] -2.028887e-02  2.152896e+00  1.075200e-01  7.209945e-01 -7.287453e-03
#> [396]  1.303361e+00  1.937865e-01  2.109218e+00  1.471748e+00  1.979938e+00
#> [401]  2.030087e-01  8.215309e-02  4.289506e-01  2.023363e-01  1.397064e+00
#> [406]  9.309552e-01  1.033104e+00  2.048411e+00  6.547545e-01  2.215231e+00
#> [411] -3.018351e-01  2.048130e+00 -5.712236e-03  2.120885e+00  9.082471e-02
#> [416]  2.030244e+00  3.541061e-01  1.596732e+00  9.421243e-01  2.077801e+00
#> [421]  5.621108e-01  1.956013e+00 -9.291621e-03  2.083575e+00  2.074809e-02
#> [426]  1.793689e+00  1.592781e+00  5.716886e-01  3.950777e-02  2.081762e+00
#> [431]  1.869709e-01  2.100695e+00  1.713598e+00  1.857348e+00  2.078797e-01
#> [436]  2.074368e+00  3.009729e-01  1.832266e+00  6.871936e-01  1.977177e+00
#> [441]  2.368392e-01  1.658820e+00  6.392240e-01  2.119992e+00  4.642977e-01
#> [446]  2.097653e+00  1.771025e-01  1.926160e+00  8.133448e-02  1.390929e+00
#> [451]  4.254299e-01  2.018279e+00 -5.362761e-02  2.098179e+00  2.224289e-01
#> [456]  2.194128e+00  3.687050e-02  2.105787e+00  8.954157e-02  1.168497e+00
#> [461]  1.064203e+00  1.663929e+00 -1.073958e-01  2.105417e+00  1.209682e-01
#> [466]  2.111965e+00  1.212615e-01  1.959717e+00  1.850217e-01  2.044225e+00
#> [471]  2.795556e-01  1.293072e+00 -7.492686e-02  1.150564e+00 -5.266211e-01
#> [476]  1.913254e+00 -3.921959e-02  1.924006e+00  2.899878e-01  1.935637e+00
#> [481]  1.202860e-01  1.980266e+00 -2.310741e-01  2.004419e+00  9.361489e-03
#> [486]  8.598171e-01 -4.650421e-03  1.318610e+00 -8.166525e-02  1.465269e+00
#> [491] -1.737787e-01  1.479858e+00  5.044172e-02  8.483030e-01  1.628890e+00
#> [496]  2.035530e+00 -3.357204e-01  2.125605e+00  3.084229e-01  1.673312e+00

x2 = makeobject(data, pi0 = 0.5, method = "nptll") # equivalent to normal
r2 = computemixdist(x2)
posteriormean(data, r2, fun = function(x) x^2)

#>   [1]  0.1814562  2.5643257  0.3636249  4.1709409  0.4842833  0.9927928
#>   [7]  0.4433489  3.6600767  0.4076515  4.3694212  0.3552920  5.3340039
#>  [13]  0.4628415  4.9787377  2.1166498  2.0487450  0.2130592  3.7670871
#>  [19]  0.2224103  4.2000880  3.6738266  1.2692568  1.0872003  4.2848183
#>  [25]  0.1801921  2.3832155  0.2652988  3.9271879  0.6807840  4.4568421
#>  [31]  0.1908430  1.7890744  0.5724977  4.0263840  0.1860753  4.1900723
#>  [37]  1.5842516  4.4171768  0.7596407  3.6982195  0.2565945  4.3022119
#>  [43]  0.2813027  4.3062236  0.2407766  0.8711169  0.4373767  4.8683402
#>  [49]  2.5173986  4.3307688  1.7699978  4.2328466  0.1785024  4.2405821
#>  [55]  1.9345241  1.7505263  0.2921887  4.1071033  1.6138085  3.9799689
#>  [61]  0.3219692  3.4365005  0.6415149  4.3541089  3.8394979  3.9277255
#>  [67]  2.1884060  4.3947249  0.2034769  3.7212660  1.1926398  4.0187714
#>  [73]  0.3475498  4.0864745  0.3411424  2.6509945  0.1790270  1.3107987
#>  [79]  0.2204866  2.7582846  0.5990099  2.8614299  0.2055529  4.3900786
#>  [85]  0.4990020  5.5646393  3.1927424  2.4355815  0.7842988  3.0929139
#>  [91]  0.2186635  3.6588705  0.5063876  3.3564497  2.9007864  3.8741715
#>  [97]  0.2107681  1.8584731  2.4182564  0.7425054  0.2867634  4.0162795
#> [103]  0.6777119  4.1016055  0.6026329  3.3881039  0.9532167  4.3861315
#> [109]  0.2427092  4.1573881  1.6878121  3.1447722  1.2380821  0.7663689
#> [115]  1.5559792  3.4837220  0.6327064  4.9436628  0.2858139  3.6567600
#> [121]  0.1914197  5.0931522  0.3299622  4.0928653  0.6812177  3.4140885
#> [127]  0.2011029  4.0403630  0.3072289  4.2104378  0.1985885  4.4328557
#> [133]  1.8870755  1.5882893  3.7472798  4.0012586  2.2781387  1.0630728
#> [139]  3.1121353  4.2458445  0.3255195  2.7357520  0.7446820  1.9706365
#> [145]  4.2619148  3.5634907  0.2204022  3.3007579  1.1617175  2.9048583
#> [151]  0.2190496  4.7723079  0.1976946  3.1480531  0.6907386  4.3321604
#> [157]  1.7640176  4.3133650  0.2312333  3.2255046  0.4770375  2.1558148
#> [163]  3.2166869  4.0564328  0.3258219  1.3943953  0.2277515  0.6026147
#> [169]  0.3373537  4.2514851  0.1855615  4.4433519  0.1781740  1.4052404
#> [175]  0.4440056  2.6032151  0.4289885  3.0650744  3.5636925  4.3218027
#> [181]  2.1490588  4.4312644  0.2487998 11.7146271  0.2792998  3.3463564
#> [187]  0.3681005  3.6544306  0.1906602  4.5349492  0.3489441  1.1240531
#> [193]  0.3634732  4.3324442  0.3988305  2.3447182  0.8064676  7.4438857
#> [199]  2.1701226  4.2637590  0.6757242  3.8971594  1.9852421  3.5413871
#> [205]  3.7622502  3.8315929  1.1757076  4.3681111  0.2652578  4.3592859
#> [211]  0.3686665  0.6346251  0.1816421  4.4492629  3.7827793  3.0162749
#> [217]  1.0117569  3.4354030  0.1784853  2.2297869  1.1792633  4.4205000
#> [223]  0.8803288  4.3099312  0.1783589  2.7820723  0.1816680  4.7663735
#> [229]  4.4294015  4.1202463  1.1352825  3.9487789  0.1915612  4.5018450
#> [235]  0.4615564  4.3748210  0.1794299  4.4485096  1.7572295  4.8360767
#> [241]  0.4594644  4.3376891  0.2356846  4.1558364  0.7381098  2.5116925
#> [247]  0.9078263  4.2475174  0.2942349  3.9517836  0.1928910  4.6898346
#> [253]  0.6203820  4.6484726  0.6545676  4.4401310  2.4000311  3.1322373
#> [259]  2.2332492  2.9992894  0.9708258  4.6688105  3.6505752  3.1984979
#> [265]  3.9234231  2.6805413  0.2557215  4.4121141  0.3199249  4.0852161
#> [271]  0.2067333  4.1412028  0.4353132  1.7822706  0.3805288  2.3002000
#> [277]  0.2471236  5.3387133  1.1512317  1.1551087  0.3835149  3.1187781
#> [283]  0.3366646  3.7254636  0.3084235  2.3952727  0.4114830  1.6599019
#> [289]  0.5327601  1.6934753  0.4356051  2.5383142  0.2468851  4.2991404
#> [295]  0.5806262  3.5639616  0.1913897  1.9861275  2.1032068  4.3841884
#> [301]  0.2028622  5.6026965  0.2747247  9.5234185  0.7768347  4.2046071
#> [307]  0.2357052  4.0302112  1.8695388  2.0490473  1.0736857  2.8830871
#> [313]  0.5280921  0.3756971  0.1785152  4.6285206  0.3588485  4.1952316
#> [319]  0.1865188  3.7865828  0.1791981  1.3173681  3.0866032  2.6889721
#> [325]  0.1788011  2.1035516  0.2054337  4.6299721  0.4462011  1.7084803
#> [331]  0.7082950  3.8983847  0.1902706  4.3626132  0.1811178  4.3649463
#> [337]  2.3017476  4.3122429  0.2409903  4.4975858  0.1829360  3.6333470
#> [343]  0.1805395  0.5887537  2.5025866  3.9050075  0.3197209  2.9766657
#> [349]  2.4211452  4.0371383  0.3211980  2.7881185  0.1792012  4.2078644
#> [355]  1.2886915  6.5399687  3.9846118  6.0103979  2.6853921  3.3143039
#> [361]  0.3695578  2.8295961  0.1807025  3.4394393  0.1954841  2.6581579
#> [367]  3.0868645  4.3053193  0.1834219  0.8462051  2.3722396  4.2311845
#> [373]  0.5056090  4.4662906  1.9720929  3.6727943  0.1915413  4.1256013
#> [379]  0.9475068  4.7664377  1.2613923  2.8736901  1.3683767  4.4193964
#> [385]  2.5724845  4.0695495  0.3781082  3.7936970  0.1814593  0.5027286
#> [391]  0.1837935  4.7806753  0.3307864  1.5366174  0.1905665  2.7478686
#> [397]  0.4813973  4.5551184  3.1006760  4.1830602  0.4984484  0.2912683
#> [403]  0.9401043  0.4972006  2.9440686  1.9713246  2.1837576  4.3481933
#> [409]  1.4002222  5.1523400  0.4265860  4.3474422  0.1915578  4.6097480
#> [415]  0.3044260  4.3014421  0.7904343  3.3632997  1.9945272  4.4345120
#> [421]  1.2103429  4.1294962  0.1893574  4.4539661  0.2132472  3.7794401
#> [427]  3.3549853  1.2299162  0.2336394  4.4477426  0.4688838  4.5186020
#> [433]  3.6097281  3.9152309  0.5075061  4.4234221  0.6857374  3.8615973
#> [439]  1.4669612  4.1768193  0.5620064  3.4940871  1.3683124  4.6053894
#> [445]  1.0114355  4.5062860  0.4509094  4.0638499  0.2900472  2.9312120
#> [451]  0.9330193  4.2721846  0.1784616  4.5083901  0.5347558  5.0238027
#> [457]  0.2305562  4.5400623  0.3024544  2.4660254  2.2485260  3.5048620
#> [463]  0.2029497  4.5384649  0.3528624  4.5675124  0.3533513  4.1377222
#> [469]  0.4653195  4.3371082  0.6440365  2.7263431  0.1838716  2.4285969
#> [475]  0.7301887  4.0357473  0.1786202  4.0591497  0.6643070  4.0845804
#> [481]  0.3517263  4.1838018  0.3355786  4.2393162  0.2027995  1.8237069
#> [487]  0.1922463  2.7797755  0.1868560  3.0870818  0.2671870  3.1176957
#> [493]  0.2470962  1.7998427  3.4310075  4.3147095  0.4714416  4.6332366
#> [499]  0.7003177  3.5246544

data = runif(500, min = -0.5, max = 0.5)
x3 = makeobject(data, method = "npnormcll", beta = 100)
r3 = computemixdist(x3)
posteriormean(data, r3)

#>   [1] -0.112589061  0.374296900 -0.188949270 -0.086489207 -0.174412464
#>   [6]  0.356576320  0.294635891  0.016543591  0.348894030  0.371812625
#>  [11]  0.372522747 -0.053126835  0.374119210  0.225490708 -0.374191944
#>  [16]  0.127437669 -0.083978161  0.329801518  0.149036519 -0.382515462
#>  [21] -0.080746591 -0.036262509 -0.068982751 -0.273273330  0.364591723
#>  [26]  0.029098945 -0.135015338  0.278674767  0.371684213  0.220422360
#>  [31] -0.150224590  0.026141950  0.214018759 -0.313137906 -0.382664023
#>  [36]  0.134611348  0.253492380  0.003024923 -0.252324995  0.202792734
#>  [41]  0.205237937 -0.365757457  0.028183650  0.111882025 -0.234773370
#>  [46] -0.067149863 -0.163063661 -0.015956381  0.299046847 -0.224198065
#>  [51] -0.328121362  0.333462256  0.326977602 -0.026329388  0.195451821
#>  [56]  0.127509012  0.103585684  0.119719711 -0.107619495  0.112400085
#>  [61] -0.003672050  0.309505372  0.099619995  0.360233868  0.180443038
#>  [66] -0.070163646  0.360223569 -0.039623127 -0.361675647  0.327343290
#>  [71]  0.303731942 -0.048773225 -0.024477569  0.069902263  0.274495298
#>  [76]  0.049747619  0.373113023  0.083815968  0.333422682 -0.049003189
#>  [81] -0.062475060  0.323973346 -0.224892149  0.312514086 -0.378395749
#>  [86]  0.255482499 -0.074101376  0.373393862 -0.229906516 -0.325506516
#>  [91]  0.357300126 -0.170436308 -0.196248516  0.147967760  0.155592866
#>  [96] -0.029325425 -0.377856302  0.259768460  0.011607080 -0.131400944
#> [101] -0.053121914  0.009622172 -0.188072692  0.371019711  0.364827579
#> [106]  0.307200468 -0.037762876 -0.052713866 -0.336958189  0.222439722
#> [111] -0.135079713 -0.025656092  0.241809224 -0.175500008  0.218725079
#> [116]  0.085828582  0.371876080 -0.082171534 -0.345589245 -0.166156160
#> [121] -0.132974335 -0.283270458  0.090110253  0.370164186 -0.229679560
#> [126] -0.290623172  0.348258427 -0.017958608 -0.197600681  0.348276898
#> [131]  0.342134746 -0.365552807  0.373249620  0.008856125  0.008806684
#> [136] -0.293219657  0.076196496 -0.156401419 -0.317805411  0.292701485
#> [141] -0.040544411 -0.375651146 -0.346330588 -0.379771626 -0.115480387
#> [146] -0.220533009  0.102539571 -0.224742996 -0.176084370 -0.203151078
#> [151] -0.344525091  0.048072035 -0.187512832 -0.064122854  0.086459974
#> [156] -0.179057575  0.349532734 -0.384855591  0.270532519 -0.140300705
#> [161] -0.052534153 -0.020057919 -0.053307411  0.213342153 -0.091877742
#> [166] -0.384629791  0.370904318 -0.062662924 -0.134979593 -0.381034617
#> [171] -0.248930003 -0.359761620  0.337222181  0.028398808 -0.146154561
#> [176]  0.027246454  0.170200907 -0.050764984  0.218512789  0.228318119
#> [181] -0.312330376  0.369810553 -0.048289318  0.348529428 -0.382680061
#> [186]  0.226428975  0.128906524  0.094378235  0.201974013 -0.365219744
#> [191]  0.355468653  0.188588565  0.159541285 -0.175115445  0.297746215
#> [196]  0.368823488 -0.012078503 -0.236522682 -0.027424864 -0.021950413
#> [201]  0.074078158 -0.238524426  0.194826892 -0.187978802  0.272122964
#> [206]  0.372146924  0.245468201 -0.025271043  0.227164534 -0.005363867
#> [211]  0.127356230 -0.354759129  0.372614557  0.373555450 -0.345537543
#> [216] -0.384988399 -0.163796302  0.109351574 -0.117593340 -0.302837442
#> [221]  0.369585041  0.206783072 -0.367957478  0.345178882  0.300907434
#> [226]  0.249046635  0.126623423 -0.312138224  0.359469195 -0.383175230
#> [231] -0.092615250  0.368382124 -0.218374160  0.372529868 -0.035204186
#> [236] -0.051235838 -0.365576748 -0.035860384  0.236513881  0.011871682
#> [241] -0.058292584  0.282379815 -0.211137721 -0.010836998 -0.012900893
#> [246]  0.374768643  0.181215221 -0.001983355 -0.031014967  0.370837491
#> [251]  0.062473005 -0.185638022  0.366988793 -0.187394612 -0.046741685
#> [256] -0.151380563  0.246540714 -0.067529230 -0.128459610  0.290936828
#> [261] -0.036939519 -0.253162919 -0.335782520  0.365363013 -0.380756299
#> [266] -0.215068010  0.367203033  0.112482994 -0.132519014  0.325418864
#> [271] -0.107108106  0.338141117  0.366058360 -0.019495735 -0.045545513
#> [276] -0.017219118  0.342223250 -0.368102064 -0.233799740  0.370130955
#> [281]  0.375115964  0.151451276 -0.258737204  0.374718731 -0.339429705
#> [286]  0.024127235 -0.109323785  0.372688464 -0.033004423 -0.361620302
#> [291] -0.099431333 -0.327195681  0.158740014  0.092033825 -0.067471001
#> [296]  0.013246167 -0.384946051  0.325451126 -0.308260414  0.348296359
#> [301]  0.307322424 -0.079685881 -0.365226601 -0.053357945  0.259321038
#> [306] -0.144881326 -0.221751193  0.319073518  0.373617585 -0.047692023
#> [311]  0.369011849  0.362500257 -0.378124779  0.199521741  0.375100497
#> [316]  0.233750920  0.042567686 -0.384449321  0.022567328  0.009686059
#> [321] -0.136379353  0.101232228  0.035343397  0.373983464 -0.146078758
#> [326]  0.040468758 -0.359157528  0.274325863 -0.066533191 -0.380932580
#> [331] -0.293146710 -0.018509433 -0.352832256  0.324529684  0.303253841
#> [336] -0.385401173 -0.226877666 -0.306226056  0.190446074  0.028748587
#> [341] -0.136563859 -0.209474865  0.029114171  0.206440587 -0.003499302
#> [346] -0.382699199  0.348573256 -0.134051245  0.371613190  0.328191988
#> [351]  0.043532506 -0.331939931 -0.100564800  0.040087537  0.042487615
#> [356] -0.293566559 -0.092639217 -0.266944894  0.215820938 -0.002326220
#> [361]  0.161110843  0.367370749 -0.313446310 -0.378909802 -0.385324065
#> [366]  0.192174026  0.350236474 -0.037509676 -0.037216480 -0.385513159
#> [371]  0.158892819 -0.005621424  0.324887931  0.346705756 -0.029575871
#> [376] -0.382466942 -0.195221961 -0.328384399 -0.031765309  0.342819851
#> [381]  0.183257658  0.330343274 -0.382723421  0.373748554 -0.068215777
#> [386] -0.371013649 -0.047299992 -0.043684701 -0.208089729 -0.271439092
#> [391] -0.269282155 -0.175661486  0.179264566  0.134167775 -0.375653338
#> [396]  0.074379238 -0.109865298 -0.119633857 -0.268759090  0.360497515
#> [401]  0.319358755 -0.066971687 -0.034491846  0.369901743  0.366085099
#> [406]  0.099721662 -0.376606272 -0.385691728 -0.190889430  0.370357623
#> [411] -0.377517291  0.374625430  0.184237195 -0.139990206  0.373666946
#> [416] -0.021427374  0.266596489 -0.306720029  0.060539187  0.374414415
#> [421]  0.274011185  0.187119559 -0.028462606  0.333824942  0.043950834
#> [426]  0.364664584  0.240280791 -0.384024066 -0.004188890 -0.317462153
#> [431] -0.023189208 -0.084790306  0.178219045 -0.012588195  0.003368997
#> [436] -0.003284992 -0.191123945 -0.048345448 -0.114045456  0.110313068
#> [441]  0.344506902 -0.260407785 -0.138416160 -0.006191036 -0.037509824
#> [446] -0.210207145 -0.195067446  0.257052943 -0.038258439 -0.229534936
#> [451]  0.124041164  0.347857933 -0.160578467 -0.359606475 -0.287646091
#> [456] -0.377409223  0.012216210  0.228578509 -0.370723446  0.323401387
#> [461]  0.333297869  0.150483649  0.085135741 -0.111192569 -0.346072026
#> [466]  0.231959321  0.109257352 -0.307625503 -0.382474100 -0.163554727
#> [471]  0.372315813  0.260929784 -0.385324600 -0.228850997  0.143540146
#> [476] -0.380542921  0.344028784  0.062717350  0.374196555 -0.134060469
#> [481] -0.018251995 -0.085316925 -0.019438818  0.309417705  0.256486941
#> [486] -0.132213983 -0.044481721  0.331001499 -0.368567994  0.008459638
#> [491]  0.139030350  0.260623826 -0.340379571 -0.071307039  0.231419705
#> [496] -0.155636090  0.243216821 -0.033593023 -0.064130980 -0.382588355